Add files via upload
This zip folder contains a comprehensive collection of all the essential files required for the successful execution of the code. It includes:

-.txt Files: These text files may contain configuration settings, input data, documentation, or other crucial information that the program needs during runtime.
-.c Files: These are the source code files written in the C programming language. They are vital for compiling and building the executable program.

By providing both the .txt and .c files, this package ensures that you have all the components necessary to compile, configure, and run the code without any missing dependencies.